+++ include/LightGBM/utils/random.h
@@ -52,12 +52,12 @@ public:
   * \param K
   * \return K Ordered sampled data from {0,1,...,N-1}
   */
-  inline std::vector<size_t> Sample(size_t N, size_t K) {
-    std::vector<size_t> ret;
+  inline std::vector<int> Sample(int N, int K) {
+    std::vector<int> ret;
     if (K > N || K < 0) {
       return ret;
     }
-    for (size_t i = 0; i < N; ++i) {
+    for (int i = 0; i < N; ++i) {
       double prob = (K - ret.size()) / static_cast<double>(N - i);
       if (NextDouble() < prob) {
         ret.push_back(i);
